requirements.txt
APScheduler==3.0.3
pyasn1==0.1.7
pyasn1-modules==0.0.5
pytz==2015.4
PyYAML==3.11
requests==2.7.0
six==1.9.0
sleekxmpp==1.3.1
typing==3.5.0b1
tzlocal==1.1.3
websocket-client==0.32.0
